Deleting a board now deletes all its pins
