+++ src/javascripts/helpers/views/startApp.js
@@ -3,6 +3,7 @@ import domBuilder from '../../components/domBuilder';
 import logoutButton from '../../components/logoutButton';
 import navBar from '../../components/navBar';
 import { getBoards } from '../data/boardData';
+import domEvents from '../events/domEvents';
 import navigationEvents from '../events/navigationEvents';
 import showTitle from './boardsTitle';
 
@@ -11,6 +12,7 @@ const startApp = (user) => {
   navBar();
   logoutButton();
   navigationEvents();
+  domEvents();
   showTitle();
   getBoards().then((boards) => showBoards(boards));
 };
